requirements.txt
affine==2.4.0
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
contourpy==1.2.0
cycler==0.12.1
dataclasses-json==0.6.4
distro==1.9.0
fastapi==0.110.0
fiona==1.9.6
fonttools==4.50.0
frozenlist==1.4.1
geopandas==0.14.3
h11==0.14.0
h3==3.7.7
httpcore==1.0.4
httpx==0.27.0
idna==3.6
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
kiwisolver==1.4.5
langsmith==0.1.27
marshmallow==3.21.1
matplotlib==3.8.3
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.2.1
numpy==1.26.4
openai==1.14.1
orjson==3.9.15
osmnx==1.9.1
packaging==23.2
pandas==2.2.1
pillow==10.2.0
pydantic==2.6.4
pydantic_core==2.16.3
pyparsing==3.1.2
pyproj==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
rasterio==1.3.9
requests==2.31.0
rioxarray==0.15.1
scikit-learn==1.4.1.post1
scipy==1.12.0
setuptools==68.2.2
shapely==2.0.3
six==1.16.0
sniffio==1.3.1
snuggs==1.4.7
SQLAlchemy==2.0.28
starlette==0.36.3
tenacity==8.2.3
termcolor==2.4.0
threadpoolctl==3.3.0
tqdm==4.66.2
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.28.1
wheel==0.41.2
xarray==2024.2.0
yarl==1.9.4
earthengine-api==0.1.313